New plugin
This plugin is extremely light weight and it basically just whitelists
a bunch of different apps for us to be able to interact with. fb: links
will work. This is required because even though we go
window.open(“facebook.com”, “_system”); iPhones automatically turn it
into an fb:// link so we need permission for it to open that link